Move temp_env to a test/utils.py file and use it in test_client.py
