+++ tests/test_client.py
@@ -1,9 +1,11 @@
+import os
 import sys
 import tempfile
 
 from pytest_httpserver import HTTPServer
 
 from denvr.client import client
+from tests.utils import temp_env
 
 
 def test_client(httpserver: HTTPServer):
@@ -28,21 +30,16 @@ def test_client(httpserver: HTTPServer):
     content = """
     [defaults]
     server = "{}"
-    api = "v2"
-    cluster = "Hou1"
-    tenant = "denvr"
-    vpcid = "denvr"
-    rpool = "reserved-denvr"
-    retries = 5
-
-    [credentials]
-    username = "alice@denvrtest.com"
-    password = "alice.is.the.best"
     """.format(httpserver.url_for("/"))
     kwargs = {"delete_on_close": False} if sys.version_info >= (3, 12) else {"delete": False}
     with tempfile.NamedTemporaryFile(**kwargs) as fp:
         fp.write(content.encode())
         fp.close()
 
-        virtual = client("servers/virtual")
-        assert type(virtual).__name__ == "Client"
+        with temp_env():
+            os.environ["DENVR_CONFIG"] = fp.name
+            os.environ["DENVR_USERNAME"] = "alice@denvrtest.com"
+            os.environ["DENVR_PASSWORD"] = "alice.is.the.best"
+
+            virtual = client("servers/virtual")
+            assert type(virtual).__name__ == "Client"
